ci rename package
