+++ src/main/java/com/example/baeldungwebapp/BaeldungWebApplication.java
@@ -1,13 +1,13 @@
-package com.example.demo;
+package com.example.baeldungwebapp;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class DemoApplication {
+public class BaeldungWebApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(DemoApplication.class, args);
+		SpringApplication.run(BaeldungWebApplication.class, args);
 	}
 
 }
