+++ src/test/java/com/example/baeldungwebapp/BaeldungWebApplicationTests.java
@@ -1,10 +1,10 @@
-package com.example.demo;
+package com.example.baeldungwebapp;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class DemoApplicationTests {
+class BaeldungWebApplicationTests {
 
 	@Test
 	void contextLoads() {
